feat: Add search feature

+++ app/src/main/java/com/murerwa/murerwaweather/domain/models/forecast/WeatherForecast.kt
@@ -5,7 +5,7 @@ data class WeatherForecast(
     val dt: Int,
     val dt_txt: String,
     val main: Main,
-    val pop: Int,
+    val pop: Double,
     val sys: Sys,
     val visibility: Int,
     val weather: List<Weather>,
